Change Schedule link to be new custom link
This link will allow us to not care about remembering to include
localePath and will correctly add the is-active class to link,
regardless of the query or hash

This change will be added to other links later, to avoid
breaking anything major right before a push. This is the only
page that will really benefit now as is.